requirements.txt
pycryptodome==3.15.0
requests==2.25.1
matplotlib==3.5.0
ddddocr==1.4.3

